+++ docs/source/changelog.rst
@@ -11,6 +11,15 @@ Changelog
 History
 ==========================
 
+Version 0.2.28 (2022-09-21)
+---------------------------
+
+* Fixed bug in Node repr if operator is None.
+* Added new ``population_growth_function`` to Island, easier to switch between stead-state and generational.
+* Added new population growth functions to ``utils``.
+* ``IslandError`` added to ``utils``.
+
+
 Version 0.2.27 (2022-09-01)
 ---------------------------